+++ includes/class-autoload.inc.php
@@ -1,0 +1,9 @@
+<?php
+
+spl_autoload_register('myAutoLoader');
+
+function myAutoLoader($className) {
+
+    require_once "../classes/" . $className . ".class.php";
+
+}
