+++ includes/logout.inc.php
@@ -1,0 +1,6 @@
+<?php
+
+require_once 'class-autoload.inc.php';
+
+$disconnect = new User;
+$disconnect -> disconnect();
