+++ public/css/style.css
@@ -1,12 +1,38 @@
+main
+{
+    display: flex;
+    flex-direction: column;
+    flex-wrap: wrap;
+    align-content:center;
+}
+
+header ul
+{
+    display: flex;
+    background-color: darkgoldenrod;
+    list-style-type: none;
+}
+
+header ul li
+{
+    margin: 50px;
+}
+
 form
 {
+    width: 70%;
     text-align:center;
     padding: 16px;
 }
 
+legend
+{
+    font-size: 2em;
+}
+
 label
 {
-    font-size:0.9em;
+    font-size:1.1em;
     margin-bottom:0px;
 }
 
@@ -51,125 +77,3 @@ input:valid+span:after {
     content: '✓';
     padding-left: 5px;
 }
-/*******************************Calendar Top Navigation*********************************/
-div#calendar{
-  margin:0px auto;
-  padding:0px;
-  width: 602px;
-  font-family:Helvetica, "Times New Roman", Times, serif;
-}
- 
-div#calendar div.box{
-    position:relative;
-    top:0px;
-    left:0px;
-    width:100%;
-    height:40px;
-    background-color:   #787878 ;      
-}
- 
-div#calendar div.banner{
-    line-height:40px;  
-    vertical-align:middle;
-    position:absolute;
-    left:11px;
-    top:0px;
-    width:582px;
-    height:40px;   
-    text-align:center;
-}
- 
-div#calendar div.banner a.prev,div#calendar div.banner a.next{ 
-    position:absolute;
-    top:0px;   
-    height: 17px;
-    display:block;
-    cursor:pointer;
-    text-decoration:none;
-    color:#FFF;
-}
- 
-div#calendar div.banner span.title{
-    color:#FFF;
-    font-size:18px;
-}
- 
- 
-div#calendar div.banner a.prev{
-    left:0px;
-}
- 
-div#calendar div.banner a.next{
-    right:0px;
-}
- 
- 
- 
- 
-/*******************************Calendar Content Cells*********************************/
-div#calendar div.box-content{
-    border:1px solid #787878 ;
-    border-top:none;
-}
- 
- 
- 
-div#calendar ul.label{
-    float:left;
-    margin: 0px;
-    padding: 0px;
-    margin-top:5px;
-    margin-left: 5px;
-}
- 
-div#calendar ul.label li{
-    margin:0px;
-    padding:0px;
-    margin-right:5px;  
-    float:left;
-    list-style-type:none;
-    width:80px;
-    height:40px;
-    line-height:40px;
-    vertical-align:middle;
-    text-align:center;
-    color:#000;
-    font-size: 15px;
-    background-color: transparent;
-}
- 
- 
-div#calendar ul.dates{
-    float:left;
-    margin: 0px;
-    padding: 0px;
-    margin-left: 5px;
-    margin-bottom: 5px;
-}
- 
-/** overall width = width+padding-right**/
-div#calendar ul.dates li{
-    margin:0px;
-    padding:0px;
-    margin-right:5px;
-    margin-top: 5px;
-    line-height:80px;
-    vertical-align:middle;
-    float:left;
-    list-style-type:none;
-    width:80px;
-    height:80px;
-    font-size:25px;
-    background-color: #DDD;
-    color:#000;
-    text-align:center; 
-}
- 
-:focus{
-    outline:none;
-}
- 
-div.clear{
-    clear:both;
-}
-
