--- view/errorView.php
@@ -1,1 +1,0 @@
-<!-- en préparation -->
